backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
openai==1.63.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
asyncpg==0.29.0
